split demos and start them open
