+++ packages/design-system/src/components/floating-panel/popover.stories.tsx
@@ -1,12 +1,10 @@
 import { Button } from "../button";
-import { StorySection } from "../storybook";
 import { css, theme } from "../../stitches.config";
 import { typography } from "../typography";
-import * as Dialog from "./dialog";
 import * as Popover from "./popover";
 
 export default {
-  title: "Library/Floating Panel",
+  title: "Library/Floating Panel/Popover",
 };
 
 const bodyStyles = css({
@@ -24,36 +22,8 @@ const buttonsStyle = css({
   justifyContent: "flex-end",
 });
 
-const DialogDemo = () => (
-  <Dialog.Root>
-    <Dialog.Trigger asChild>
-      <Button>Open</Button>
-    </Dialog.Trigger>
-    <Dialog.Content>
-      <div className={bodyStyles()}>
-        <Dialog.Description asChild>
-          <p className={descriptionStyles()}>This is a description</p>
-        </Dialog.Description>
-        <div className={buttonsStyle()}>
-          <Dialog.Close asChild>
-            <Button color="ghost">Cancel</Button>
-          </Dialog.Close>
-          <Dialog.Close asChild>
-            <Button color="positive">Save</Button>
-          </Dialog.Close>
-        </div>
-      </div>
-
-      {/* Title is at the end intentionally,
-       * to make the close button last in the tab order
-       */}
-      <Dialog.Title>Title</Dialog.Title>
-    </Dialog.Content>
-  </Dialog.Root>
-);
-
 const PopoverDemo = () => (
-  <Popover.Root>
+  <Popover.Root defaultOpen>
     <Popover.Trigger asChild>
       <Button>Open</Button>
     </Popover.Trigger>
@@ -77,14 +47,4 @@ const PopoverDemo = () => (
     </Popover.Content>
   </Popover.Root>
 );
-
-export const FloatingPanel = () => (
-  <>
-    <StorySection title="Dialog">
-      <DialogDemo />
-    </StorySection>
-    <StorySection title="Popover">
-      <PopoverDemo />
-    </StorySection>
-  </>
-);
+export { PopoverDemo as Popover };
